refactor: change the keyboard generation
